refactor(sdk): rename wallet-account-alignment to alignment
Shorter import path; module already lives under utils/.
